Decouple frontend UI root subproject from frontend UI table implementation
