+++ binding/config/checkstyle/checkstyle.xml
@@ -1,0 +1,1 @@
+../../../config/checkstyle/checkstyle_allow_java_collections.xml
