--- branchLayoutImpl/config/checkstyle/checkstyle.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE module PUBLIC
-        "-//Checkstyle//DTD Checkstyle Configuration 1.3//EN"
-        "https://checkstyle.org/dtds/configuration_1_3.dtd" [
-        <!ENTITY checkerModuleDirectives SYSTEM "../../../config/checkstyle/checker_module_directives.xml">
-        <!ENTITY treeWalkerModuleDirectives SYSTEM "../../../config/checkstyle/tree_walker_module_directives.xml">
-        <!ENTITY prohibitJavaCollections SYSTEM "../../../config/checkstyle/prohibit_java_collections.xml">
-        ]>
-
-<module name="Checker">
-    &checkerModuleDirectives;
-
-    <module name="TreeWalker">
-        &treeWalkerModuleDirectives;
-        &prohibitJavaCollections;
-    </module>
-</module>
