+++ frontendUiApi/build.gradle
@@ -1,0 +1,8 @@
+vavr()
+
+apply plugin: 'org.jetbrains.intellij'
+disableIntellijTasks()
+intellij {
+  version intellijVersion
+  plugins 'git4idea'
+}
