+++ frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/selection/IGitRepositorySelectionChangeObserver.java
@@ -1,0 +1,6 @@
+package com.virtuslab.gitmachete.frontend.ui.api.selection;
+
+@FunctionalInterface
+public interface IGitRepositorySelectionChangeObserver {
+  void onSelectionChanged();
+}
