+++ frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/table/IGraphTableManagerFactory.java
@@ -1,0 +1,9 @@
+package com.virtuslab.gitmachete.frontend.ui.api.table;
+
+import com.intellij.openapi.project.Project;
+
+import com.virtuslab.gitmachete.frontend.ui.api.selection.IGitRepositorySelectionProvider;
+
+public interface IGraphTableManagerFactory {
+  IGraphTableManager create(Project project, IGitRepositorySelectionProvider gitRepositorySelectionProvider);
+}
