--- frontendUiRoot/build.gradle
@@ -1,14 +1,0 @@
-vavr()
-
-dependencies {
-  implementation project(':logging')
-  implementation project(':frontendActionIds')
-  implementation project(':frontendUiTable')
-}
-
-apply plugin: 'org.jetbrains.intellij'
-disableIntellijTasks()
-intellij {
-  version intellijVersion
-  plugins 'git4idea'
-}
