--- frontendUiTable/build.gradle
@@ -1,18 +1,0 @@
-vavr()
-
-dependencies {
-  implementation project(':logging')
-  implementation project(':binding')
-  implementation project(':branchLayoutApi')
-  implementation project(':backendApi')
-  implementation project(':frontendActionIds')
-  implementation project(':frontendDataKeys')
-  implementation project(':frontendGraphApi')
-}
-
-apply plugin: 'org.jetbrains.intellij'
-disableIntellijTasks()
-intellij {
-  version intellijVersion
-  plugins 'git4idea'
-}
