--- frontendUiTable/src/main/java/com/virtuslab/gitmachete/frontend/ui/selection/ISelectionChangeObservable.java
@@ -1,7 +1,0 @@
-package com.virtuslab.gitmachete.frontend.ui.selection;
-
-public interface ISelectionChangeObservable<T> {
-  T getValue();
-
-  void addObserver(ISelectionChangeObserver observer);
-}
