--- frontendUiTable/src/main/java/com/virtuslab/gitmachete/frontend/ui/selection/ISelectionChangeObserver.java
@@ -1,6 +1,0 @@
-package com.virtuslab.gitmachete.frontend.ui.selection;
-
-@FunctionalInterface
-public interface ISelectionChangeObserver {
-  void onSelectionChanged();
-}
