+++ frontendUiTableImpl/build.gradle
@@ -1,0 +1,19 @@
+vavr()
+
+dependencies {
+  implementation project(':logging')
+  implementation project(':binding')
+  implementation project(':branchLayoutApi')
+  implementation project(':backendApi')
+  implementation project(':frontendActionIds')
+  implementation project(':frontendDataKeys')
+  implementation project(':frontendGraphApi')
+  implementation project(':frontendUiApi')
+}
+
+apply plugin: 'org.jetbrains.intellij'
+disableIntellijTasks()
+intellij {
+  version intellijVersion
+  plugins 'git4idea'
+}
