+++ settings.gradle
@@ -12,5 +12,6 @@ include 'frontendActions'
 include 'frontendDataKeys'
 include 'frontendGraphApi'
 include 'frontendGraphImpl'
-include 'frontendUiTable'
-include 'frontendUiRoot'
+include 'frontendUiApi'
+include 'frontendUiTableImpl'
+include 'frontendUiRootImpl'
