+++ src/main/resources/META-INF/plugin.xml
@@ -15,8 +15,8 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <changesViewContent tabName="Git Machete"
-                            className="com.virtuslab.gitmachete.frontend.ui.root.GitMacheteContentProvider"
-                            predicateClassName="com.virtuslab.gitmachete.frontend.ui.root.GitMacheteVisibilityPredicate"/>
+                            className="com.virtuslab.gitmachete.frontend.ui.impl.root.GitMacheteContentProvider"
+                            predicateClassName="com.virtuslab.gitmachete.frontend.ui.impl.root.GitMacheteVisibilityPredicate"/>
     </extensions>
 
     <actions>
