+++ version.gradle
@@ -1,3 +1,3 @@
 ext {
-  PLUGIN_VERSION = '0.1.10'
+  PLUGIN_VERSION = '0.1.11'
 }
